Add Dart to list of Prototypes
I am the maintainer of the Dart UUID implementation, and I have been working on v6, v7, and soon v8 for Dart. I currently have v6 and v7 implemented, and working to the best of my ability without test samples and using the Go/Zig/Python outputs/implementations as reference.

Dart only has Microsecond precision, and if used in Dart2JS, it only has millisecond precision due to limitations of JS. (I used milliseconds for v6 and micro for v7).

Currently I plan to only keep it as a branch in my code and plan to keep it updated with any draft changes.